requirements.txt
anyio==3.6.2
boto3==1.25.1
botocore==1.28.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
ecdsa==0.18.0
fastapi==0.85.1
grpcio==1.50.0
grpcio-tools==1.50.0
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
jmespath==1.0.1
numpy==1.23.4
Pillow==9.2.0
protobuf==4.21.8
pyasn1==0.4.8
pydantic==1.10.2
PyMySQL==1.0.2
python-dateutil==2.8.2
python-jose==3.3.0
python-keycloak==2.6.0
python-multipart==0.0.5
qdrant-client==0.10.3
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
